+++ src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
 import NavMenu from './components/NavMenu/NavMenu';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import LandingPage from './pages/landing/Landing';
+import LandingPage from './pages/LandingPage/LandingPage';
 
 function App() {
   return (
